fix: refactoring of styles for the listings page, moved real estate filter form styles to a separate file
