+++ frontend/src/features/settingCategoryForm/settingRealtyForm/SettingPrice.tsx
@@ -1,5 +1,5 @@
 import { UseFormRegister, UseFormWatch } from 'react-hook-form';
-import { TypeSettingRealty } from '../../../widgets/settingForm/settingRealty/TypeSettingRealty';
+import { TypeSettingRealty } from '../../../widgets/settingForm/settingRealty/libr/TypeSettingRealty';
 
 interface Props {
     watch: UseFormWatch<TypeSettingRealty>;
