+++ frontend/src/pages/category/style/1024-1439-category-page.scss
@@ -21,69 +21,6 @@
                 .filters {
                     width: 100%;
                     padding: 16px 16px 24px 16px;
-                    .filter-realty-form {
-                        width: 100%;
-                        input[type='number'] {
-                            width: 42%;
-                            height: 34px;
-                            border-radius: var(--radii);
-                            padding: 10px;
-                            border: none;
-                        }
-                        input[type='submit'] {
-                            white-space: normal;
-                            height: auto;
-                        }
-
-                        .setting-price {
-                            .setting-price-adverts {
-                                display: flex;
-                                justify-content: space-between;
-                                margin-bottom: 8px;
-                                label {
-                                    padding: 8px 3.5%;
-                                }
-                            }
-                        }
-                        .checkboxes {
-                            .bathroom {
-                                .bathroom-setting {
-                                    flex-direction: column;
-                                    align-items: start;
-                                    gap: 6px;
-                                    height: auto;
-                                }
-                            }
-                        }
-                        .form-checkbox > input {
-                            position: absolute;
-                            z-index: -1;
-                            opacity: 0;
-                        }
-
-                        .form-checkbox > span::before {
-                            content: '';
-                            display: inline-block;
-                            width: 14px;
-                            height: 14px;
-                            flex-shrink: 0;
-                            flex-grow: 0;
-                            border-radius: 4px;
-                            margin: 0 5px 0 0;
-                            background-repeat: no-repeat;
-                            background-position: center center;
-                            background-size: 50% 50%;
-                            background-color: var(--grey-1);
-                        }
-
-                        .form-checkbox > input:checked + span::before {
-                            background: url('../../../shared/ui/icons/union.svg')
-                                no-repeat center center;
-                            background-size: 60%;
-                            background-color: var(--main-green);
-                            box-shadow: none;
-                        }
-                    }
                 }
             }
