+++ frontend/src/pages/category/style/768-1023-category-page.scss
@@ -42,74 +42,6 @@
                 .filters {
                     width: 100%;
                     padding: 24px 24px 12px;
-                    .filter-realty-form {
-                        display: grid;
-                        grid-template-columns: 1fr 1fr;
-                        grid-template-rows: 120px 80px 80px 70px;
-                        grid-template-areas:
-                            'price type'
-                            'area check'
-                            'liv-space check'
-                            'comission somission'
-                            'send btn-reset';
-
-                        input[type='number'] {
-                            width: 44%;
-                            height: 34px;
-                            border-radius: var(--radii);
-                            padding: 10px;
-                            border: none;
-                        }
-                        input[type='submit'] {
-                            white-space: normal;
-                            width: 82%;
-                            height: auto;
-                        }
-                        .type-of-property,
-                        .setting-price,
-                        .total-area,
-                        .living-space,
-                        .rooms-quantity,
-                        .bathroom,
-                        .balcony,
-                        .only-with-photo {
-                            width: 80%;
-                        }
-                        .realtyComission {
-                            grid-area: comission;
-                        }
-                        .type-of-property {
-                            grid-area: type;
-                        }
-                        .setting-price {
-                            grid-area: price;
-                        }
-                        .total-area {
-                            grid-area: area;
-                        }
-                        .living-space {
-                            grid-area: liv-space;
-                        }
-                        .checkboxes {
-                            grid-area: check;
-                        }
-                        input[type='submit'] {
-                            grid-area: send;
-                            width: 85%;
-                            margin-top: 10px;
-                        }
-                        .reset-setting-form {
-                            grid-area: btn-reset;
-                            width: 85%;
-                            height: 45px;
-                            margin-top: 10px;
-                            width: 82%;
-                        }
-                        .checkboxes {
-                            position: relative;
-                            bottom: 43px;
-                        }
-                    }
                 }
             }
             .adverts {
